@@ -7,7 +7,7 @@ name = "model-bakery"
 dynamic = ["version"]
 description = "Smart object creation facility for Django."
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 license = {text = "Apache License 2.0"}
 authors = [
     { name = "berin", email = "bernardoxhc@gmail.com" },
@@ -32,7 +32,6 @@ classifiers = [
     "Operating System :: OS Independent",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
