@@ -4,9 +4,7 @@ import itertools
 from typing import (
     Any,
     Generic,
-    Optional,
     TypeVar,
-    Union,
     cast,
     overload,
 )
@@ -27,7 +25,7 @@ finder = baker.ModelFinder()
 class Recipe(Generic[M]):
     _T = TypeVar("_T", bound="Recipe[M]")
 
-    def __init__(self, _model: Union[str, type[M]], **attrs: Any) -> None:
+    def __init__(self, _model: str | type[M], **attrs: Any) -> None:
         self.attr_mapping = attrs
         self._model = _model
         # _iterator_backups will hold values of the form (backup_iterator, usable_iterator).
@@ -93,7 +91,7 @@ class Recipe(Generic[M]):
         _create_files: bool = False,
         _using: str = "",
         _bulk_create: bool = False,
-        _save_kwargs: Optional[dict[str, Any]] = None,
+        _save_kwargs: dict[str, Any] | None = None,
         **attrs: Any,
     ) -> M: ...
 
@@ -106,21 +104,21 @@ class Recipe(Generic[M]):
         _create_files: bool = False,
         _using: str = "",
         _bulk_create: bool = False,
-        _save_kwargs: Optional[dict[str, Any]] = None,
+        _save_kwargs: dict[str, Any] | None = None,
         **attrs: Any,
     ) -> list[M]: ...
 
     def make(
         self,
-        _quantity: Optional[int] = None,
-        make_m2m: Optional[bool] = None,
-        _refresh_after_create: Optional[bool] = None,
-        _create_files: Optional[bool] = None,
+        _quantity: int | None = None,
+        make_m2m: bool | None = None,
+        _refresh_after_create: bool | None = None,
+        _create_files: bool | None = None,
         _using: str = "",
-        _bulk_create: Optional[bool] = None,
-        _save_kwargs: Optional[dict[str, Any]] = None,
+        _bulk_create: bool | None = None,
+        _save_kwargs: dict[str, Any] | None = None,
         **attrs: Any,
-    ) -> Union[M, list[M]]:
+    ) -> M | list[M]:
         defaults = {}
         if _quantity is not None:
             defaults["_quantity"] = _quantity
@@ -158,11 +156,11 @@ class Recipe(Generic[M]):
 
     def prepare(
         self,
-        _quantity: Optional[int] = None,
+        _quantity: int | None = None,
         _save_related: bool = False,
         _using: str = "",
         **attrs: Any,
-    ) -> Union[M, list[M]]:
+    ) -> M | list[M]:
         defaults = {
             "_save_related": _save_related,
         }
@@ -209,7 +207,7 @@ class RecipeForeignKey(Generic[M]):
 
 
 def foreign_key(
-    recipe: Union[Recipe[M], str], one_to_one: bool = False
+    recipe: Recipe[M] | str, one_to_one: bool = False
 ) -> RecipeForeignKey[M]:
     """Return a `RecipeForeignKey`.
 
@@ -232,7 +230,7 @@ def foreign_key(
 
 
 class related(Generic[M]):  # FIXME
-    def __init__(self, *args: Union[str, Recipe[M]]) -> None:
+    def __init__(self, *args: str | Recipe[M]) -> None:
         self.related = []  # type: list[Recipe[M]]
         for recipe in args:
             if isinstance(recipe, Recipe):
@@ -246,6 +244,6 @@ class related(Generic[M]):  # FIXME
             else:
                 raise TypeError("Not a recipe")
 
-    def make(self) -> list[Union[M, list[M]]]:
+    def make(self) -> list[M | list[M]]:
         """Persist objects to m2m relation."""
         return [m.make() for m in self.related]
